% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbwsdid-package.R
\docType{package}
\name{cbwsdid-package}
\alias{cbwsdid-package}
\title{cbwsdid: Covariate-Balanced Weighted Stacked DID}
\description{
Package-level documentation and imports for \code{cbwsdid}.
}
\author{
\strong{Maintainer}: Vadim Ustyuzhanin \email{vadim.v.ustyuzhanin@gmail.com} (\href{https://orcid.org/0000-0003-3800-1108}{ORCID})

}
\keyword{internal}
